requirements_original.txt
python-docx
requests
pydub
ffmpeg-python
python-dotenv
